# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MaKames Digital Business Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterUserBody {
  username: string;
  email: string;
  password: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginUserBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  referralCode: string;
  /** @nullable */
  referredBy?: number | null;
  freeDeployDaysLeft: number;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export interface Wallet {
  id: number;
  userId: number;
  balanceMd: number;
  balanceKes: number;
  updatedAt: string;
}

export type TopUpBodyPaymentMethod =
  (typeof TopUpBodyPaymentMethod)[keyof typeof TopUpBodyPaymentMethod];

export const TopUpBodyPaymentMethod = {
  mpesa: "mpesa",
  card: "card",
  international: "international",
} as const;

export interface TopUpBody {
  amountKes: number;
  paymentMethod: TopUpBodyPaymentMethod;
  paymentReference?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  topup: "topup",
  deduction: "deduction",
  refund: "refund",
  bonus: "bonus",
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amountMd: number;
  description: string;
  createdAt: string;
}

export interface BotType {
  id: string;
  name: string;
  description: string;
  costMd: number;
  apiEndpoint: string;
  features: string[];
  isActive: boolean;
}

export type BotDeploymentStatus =
  (typeof BotDeploymentStatus)[keyof typeof BotDeploymentStatus];

export const BotDeploymentStatus = {
  running: "running",
  stopped: "stopped",
  error: "error",
  pending: "pending",
} as const;

export interface BotDeployment {
  id: number;
  userId: number;
  botTypeId: string;
  botName: string;
  status: BotDeploymentStatus;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  config?: string | null;
  deployedAt: string;
  /** @nullable */
  expiresAt?: string | null;
  updatedAt: string;
}

export interface DeployBotBody {
  userId: number;
  botTypeId: string;
  botName: string;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  config?: string | null;
  useFreeDeployment?: boolean;
}

/**
 * @nullable
 */
export type UpdateDeploymentBodyStatus =
  | (typeof UpdateDeploymentBodyStatus)[keyof typeof UpdateDeploymentBodyStatus]
  | null;

export const UpdateDeploymentBodyStatus = {
  running: "running",
  stopped: "stopped",
} as const;

export interface UpdateDeploymentBody {
  /** @nullable */
  botName?: string | null;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  config?: string | null;
  /** @nullable */
  status?: UpdateDeploymentBodyStatus;
}

export interface ReferralEntry {
  id: number;
  referredUsername: string;
  joinedAt: string;
}

export interface ReferralInfo {
  userId: number;
  referralCode: string;
  referralCount: number;
  freeDeployDaysLeft: number;
  referrals: ReferralEntry[];
}

export interface ApplyReferralBody {
  userId: number;
  referralCode: string;
}

export interface ReferralResult {
  success: boolean;
  message: string;
  freeDeployDaysGranted: number;
}

export type ListDeploymentsParams = {
  userId: number;
};