# lib/api-zod/src/generated/types/applyReferralBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MaKames Digital Business Center API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplyReferralBody {
  userId: number;
  referralCode: string;
}