# lib/api-zod/src/generated/types/botDeployment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MaKames Digital Business Center API
 * OpenAPI spec version: 0.1.0
 */
import type { BotDeploymentStatus } from "./botDeploymentStatus";

export interface BotDeployment {
  id: number;
  userId: number;
  botTypeId: string;
  botName: string;
  status: BotDeploymentStatus;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  config?: string | null;
  deployedAt: Date;
  /** @nullable */
  expiresAt?: Date | null;
  updatedAt: Date;
}