# lib/api-zod/src/generated/types/referralInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MaKames Digital Business Center API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEntry } from "./referralEntry";

export interface ReferralInfo {
  userId: number;
  referralCode: string;
  referralCount: number;
  freeDeployDaysLeft: number;
  referrals: ReferralEntry[];
}