# lib/api-zod/src/generated/types/referralResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MaKames Digital Business Center API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralResult {
  success: boolean;
  message: string;
  freeDeployDaysGranted: number;
}